requirements.txt
streamlit
requests==2.27.1   
python-dotenv
diffusers==0.10.2
transformers 
scipy 
ftfy 
accelerate